+++ apps/season-editor/src/pages/components/styled-multiline-wrapper.tsx
@@ -3,5 +3,5 @@ import { styled } from "@mui/material";
 export const StyledMultilineInputWrapper = styled("div")(({ theme }) => ({
   display: "grid",
   gridTemplateColumns: "1fr 1fr 1fr",
-  gap: theme.spacing(2),
+  gap: theme.spacing(0, 2),
 }));