+++ apps/season-editor/src/pages/components/sub-competences.tsx
@@ -89,8 +89,8 @@ export const SubCompetences = ({
             key={subCompetenceField.id}
             sx={{
               display: "flex",
-              alignItems: "flex-end",
-              gap: 2,
+              alignItems: "center",
+              gap: [0, 2],
             }}
           >
             <input