+++ apps/season-editor/src/pages/types/error-object.ts
@@ -1,0 +1,4 @@
+export type ErrorObject = {
+  isError: true;
+  testItemIdDuplicates: string[];
+};