+++ apps/season-editor/src/pages/utils/index.ts
@@ -3,5 +3,6 @@ export * from "./extract-episodes";
 export * from "./extract-from-competences";
 export * from "./get-empty-season";
 export * from "./get-next-competence-id";
+export * from "./is-error-object";
 export * from "./map-to-season-object";
 export * from "./try-remove-competence";